feat(harness): add `flags` matrix mode for the rule-flag suite
Promote the rule-flag behavior suite to a first-class matrix section so it runs
from the CLI like single/transitive/idempotent, not only under `go test`.

- Move the shared suite (cases, fixture, helpers) from flag_test.go into a
  non-test flags.go so the CLI build can see it. flag_test.go keeps only the two
  go-test entry points.
- Cases now assert through a tiny `flagTB` interface (the testing.TB subset they
  use) instead of *testing.T directly. *testing.T satisfies it for `go test`;
  the CLI gets a recording shim (flagRecorder) that captures Errorf/Fatalf into
  AssertionErrors and runs Cleanups. The exact same case bodies run both ways.
- Add Matrix.ExecuteAllFlags() (CLI executor, []TestResult) and wire
  `--mode=flags` into cli/harness. flags run for --mode=flags and --mode=all;
  default stays single + idempotent. Each row reports as flags/<key> with the
  flag key in the Scenario column so the table distinguishes them.
- Docs: harness-matrix.md mode table + run examples gain the flags section;
  rule-flag-testing.md documents the flagTB / ExecuteAllFlags wiring and the CLI
  invocation. The two docs cross-link.

All 13 flags pass under both `go test -run TestRuleFlags` and
`harness matrix --mode=flags`; the registry completeness guard still holds.
